requirements.txt
# Core Deep Learning (GPU-enabled)
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Anomaly Detection
anomalib>=1.0.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0
grad-cam>=1.4.8

# ML Utilities
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Progress bars and logging
tqdm>=4.65.0

# Image handling
Pillow>=9.5.0
